fix: Extract start logic into a separate module
