+++ examples/service-clients/azure-client/external-controller/src/view.ts
@@ -6,8 +6,8 @@
 import { AzureMember, IAzureAudience } from "@fluidframework/azure-client";
 import type { LatestRaw, Presence } from "@fluidframework/presence/beta";
 
-import { ICustomUserDetails } from "./app.js";
 import { IDiceRollerController } from "./controller.js";
+import { ICustomUserDetails } from "./fluid.js";
 import type { DiceValues } from "./presence.js";
 
 function makeDiceRollerView(diceRoller: IDiceRollerController): HTMLDivElement {
